CRITICAL: seen[] was int32 while tags are int64 -- heap corruption after ~35 min
Four data-generation workers died simultaneously with munmap_chunk() / double
free after ~2150s each, with different seeds and different data. The same
elapsed time across independent processes means a fixed operation count, not a
data dependency.

Root cause: group_libs marks visited points by stamping "seen" with a
monotonically increasing int64 tag, but "seen" was allocated int32. Past 2**31
the store truncates, so "seen[q] == tag" can never be true again; the flood fill
stops marking anything visited and re-appends the same stones without bound,
running off the end of the 81-element "buf" and into the heap. At ~324 tags per
legal_mask() call that is 2**31/324 = 6.6M calls, which is about 35 minutes of
continuous play.

Fixes:
- "seen" is int64 everywhere (go.py, rollout.py, features.py). 648 bytes.
- group_libs refuses to write past the end of "buf", so a future width mismatch
  degrades into a wrong answer instead of heap corruption, which is far harder
  to trace back.
- 3 regression tests, including driving the real entry points across 2**31.
- gen_expert.py now writes a partial shard every 40 games. The old version wrote
  once at the end, so the crash destroyed 900 games of work; the worst case is
  now 40 games.

43 tests pass.

+++ ag/features.py
@@ -136,7 +136,7 @@ class FeatureExtractor:                                                         
         self.with_colour = with_colour                                                # | overwritten by the next call, and a caller stacking a batch
         self.n_planes = N_PLANES_VALUE if with_colour else N_PLANES_POLICY            # | would otherwise end up with the same position repeated.
         self.buf = np.empty(NN, dtype=np.int32)                                       # | Planes come out shaped as a square board rather than a flat
-        self.seen = np.zeros(NN, dtype=np.int32)                                      # | list because that is what a convolution consumes.
+        self.seen = np.zeros(NN, dtype=np.int64)                                      # | list because that is what a convolution consumes.
         self.tagbox = np.zeros(1, dtype=np.int64)
         self.scratch = np.zeros(NN, dtype=np.int8)
         self.out = np.zeros((self.n_planes, NN), dtype=np.float32)
